requirements.txt
addfips==0.2.2
amqp==1.4.9
anyjson==0.3.3
argh==0.26.1
asgiref==2.3.2
async-timeout==2.0.1
attrs==18.1.0
autobahn==18.6.1
Automat==0.7.0
Babel==2.2.0
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.6.0
billiard==3.3.0.23
boto3==1.6.0
botocore==1.9.0
bs4==0.0.1
CacheControl==0.12.4
celery==3.1.23
certifi==2018.11.29
channels==2.1.2
chardet==3.0.4
click==6.7
click-plugins==1.0.3
cligj==0.4.0
constantly==15.1.0
cycler==0.10.0
daphne==2.2.0
Django==1.11.13
django-annoying==0.9.0
django-celery==3.1.17
django-filter==0.13.0
django-localflavor==1.2
django-pandas==0.4.5
django-supervisor==0.3.4
djangorestframework==3.3.3
docutils==0.14
factory-boy==2.8.1
Faker==0.8.7
feedparser==5.2.1
funcsigs==1.0.2
furl==1.0.1
futures==3.0.5
geopy==1.11.0
html5lib==1.0.1
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipaddress==1.0.18
iso3166==0.8
jmespath==0.9.3
kombu==3.0.35
lxml==3.5.0
mapbox==0.15.1
Markdown==2.6.6
meld3==1.0.2
mock==2.0.0
msgpack-python==0.5.5
nltk==3.2.4
numpy==1.14.0
oauthlib==2.0.6
orderedmultidict==0.7.11
pandas==0.22.0
pathlib==1.0.1
pathtools==0.1.2
pbr==3.1.1
polyline==1.3.2
psycopg2==2.7.5
PyJWT==1.5.0
pymongo==3.2.2
pyparsing==2.1.1
python-dateutil==2.5.2
pytz==2018.4
PyYAML==3.11
redis==2.10.5
requests==2.21.0
requests-oauthlib==0.8.0
s3transfer==0.1.13
six==1.11.0
supervisor==3.2.3
text-unidecode==1.1
tornado==4.2
tweepy==3.5.0
Twisted==18.4.0
txaio==2.10.0
uritemplate==3.0.0
urllib3==1.22
watchdog==0.8.3
webencodings==0.5.1
xtermcolor==1.3
zope.interface==4.5.0
